segregation
Segregated Custom token and Custom Nft managers for readability & performance.
