+++ Packages/io.chainsafe.web3-unity.web3auth/Runtime/WalletGUI/Scripts/Web3AuthCustomNftManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4162cad38571cc4a9f0a5f5023e499c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
